transport: make SingleBand fp32-functional
SingleBand did its k-space setup (k @ Gbasis, dispersion, dE/dk) against the
shared Lattice.Gbasis, which is fp64 regardless of the working dtype, so in
fp32 the matmuls raised "float != double". Run that construction-time math in
the lattice precision and store the results into the working-dtype material
buffers; likewise build the scatter momentum-transfer magnitudes in the working
dtype. No-ops in fp64 (the default). SingleBand now runs in fp32 both ballistic
and with the e-e scatter kernel.

+++ src/qimpy/transport/material/single_band/_scatter.py
@@ -86,7 +86,9 @@ class Scatter(TreeNode):
         iq, q_index = iq_pair.flatten(0, 1).unique(dim=0, return_inverse=True)
         q = iq / kmesh
         q -= torch.round(q)  # wrap to [-0.5, 0.5) in each dimension
-        q_mag = (q @ single_band.lattice.Gbasis.T).norm(dim=-1)
+        # Gbasis is fp64 regardless of the working dtype; match it to q so the
+        # scatter matrices stay in the working precision (no-op in fp64).
+        q_mag = (q @ single_band.lattice.Gbasis.T.to(q.dtype)).norm(dim=-1)
         n_q = len(q_mag)
         log.info(f"Found {n_q} unique momentum transfers")
